Revert "Updates to multipeak capabilities"
