+++ .readthedocs.yaml
@@ -5,28 +5,24 @@
 # Required
 version: 2
 
-# Set the OS, Python version and other tools you might need
-build:
-  os: "ubuntu-22.04"
-  tools:
-    python: "3.12"
+# Set the version of Python and other tools you might need
+#build:
+#    python: "3.8"
     # You can also specify other tool versions:
-    # nodejs: "19"
-    # rust: "1.64"
-    # golang: "1.19"
+    # nodejs: "16"
+    # rust: "1.55"
+    # golang: "1.17"
 
-# Build documentation in the "docs/" directory with Sphinx
+# Build documentation in the docs/ directory with Sphinx
 sphinx:
    configuration: docs/source/conf.py
 
-# Optionally build your docs in additional formats such as PDF and ePub
+# If using Sphinx, optionally build your docs in additional formats such as PDF
 # formats:
 #    - pdf
-#    - epub
 
-# Optional but recommended, declare the Python requirements required
-# to build your documentation
-# See https://docs.readthedocs.io/en/stable/guides/reproducible-builds.html
-# python:
-#    install:
-#    - requirements: docs/requirements.txt
+# Optionally declare the Python requirements required to build your docs
+python:
+   version: 3.8
+   install:
+   - requirements: requirements.txt
