+++ build.sh
@@ -1,1 +1,1 @@
-conda install mpi4py -c conda-forge
+python setup.py install
