+++ cohere_core/controller/reconstruction_single.py
@@ -61,11 +61,7 @@ def reconstruction(lib, conf_file, datafile, dir, dev=None):
     """
     Controls single reconstruction.
 
-    This script is typically started with cohere-ui helper functions. The 'init_guess' parameter in the
-    configuration file defines whether it is a random guess, AI algorithm determined, or starting from
-    some saved state. It will set the initial guess accordingly and start phasing process. The results
-    will be saved in configured 'save_dir' parameter or in 'results_phasing' subdirectory if 'save_dir'
-    is not defined.
+    This script is typically started with cohere-ui helper functions. The 'init_guess' parameter in the configuration file defines whether it is a random guess, AI algorithm determined, or starting from some saved state. It will set the initial guess accordingly and start phasing process. The results will be saved in configured 'save_dir' parameter or in 'results_phasing' subdirectory if 'save_dir' is not defined.
 
     Parameters
     ----------
@@ -106,7 +102,7 @@ def reconstruction(lib, conf_file, datafile, dir, dev=None):
         save_dir = pars['save_dir']
     else:
         filename = conf_file.split('/')[-1]
-        save_dir = ut.join(dir, filename.replace('config_rec', 'results_phasing'))
+        save_dir = dir + '/' + filename.replace('config_rec', 'results_phasing')
 
     p = Process(target=rec_process, args=(lib, pars, datafile, dev,
                                           continue_dir, save_dir))
