+++ cohere_core/data/alien_tools.py
@@ -110,6 +110,7 @@ def analyze_clusters(arr, labels, nz):
     # print("processing labels")
     # loop over the labels (clusters).  label_counts[0] is the unique labels
     for n in range(1, nlabels):
+        #    print("   %i %i      "%(label_counts[0][n],label_counts[1][n]), end='\r')
         # the nth label from the first array of the label_counts tuple
         n_lab = label_counts[0][n]
         # the indicies of the points belonging to label n
@@ -121,6 +122,8 @@ def analyze_clusters(arr, labels, nz):
         cluster_avg[cluspts] = np.sum(arr[cluspts]) / cluspts[0].size
         # compute average asym of each cluster and store in array.
         cluster_avg_asym[cluspts] = np.sum(asymmetry[cluspts]) / cluspts[0].size
+        # print("   %i %i %f %f     "%(label_counts[0][n],label_counts[1][n],np.sum(asymmetry[cluspts]),cluspts[0].size), end='\n')
+    # print("largest clus size", cluster_size.max())
     # compute relative cluster sizes to largest (main) cluster.
     rel_cluster_size = cluster_size / cluster_size.max()
 
@@ -181,7 +184,7 @@ def save_arr(arr, dir, fname):
     """
 
     if dir is not None:
-        full_name = ut.join(dir, fname)
+        full_name = dir + '/' + fname
     else:
         full_name = fname  # save in the current dir
     tif.imsave(full_name, arr.transpose().astype(np.float32))
@@ -271,7 +274,7 @@ def auto_alien1(data, config, data_dir=None):
     if 'AA1_save_arrs' in config:
         save_arrs = config['AA1_save_arrs']
         if save_arrs:
-            save_dir = ut.join(data_dir, 'alien_analysis')
+            save_dir = data_dir + '/alien_analysis'
             if not os.path.exists(save_dir):
                 os.makedirs(save_dir)
     else:
@@ -320,7 +323,7 @@ def auto_alien1(data, config, data_dir=None):
     if (expandcleanedsig > 0):
         cuboid = np.where(cuboid > 0, 1.0, 0.0)
         sig = [expandcleanedsig, expandcleanedsig, 1.0]
-        cuboid = np.gaussian_filter(cuboid, sig)
+        cuboid = ut.gauss_conv_fft(cuboid, sig)
         no_thresh_cuboid = crop_center(data)
         cuboid = np.where(cuboid > 0.1, no_thresh_cuboid, 0.0)
     return cuboid
@@ -379,7 +382,7 @@ def filter_aliens(data, config_map):
                     return
             data = np.where((mask == 1), data, 0.0)
         else:
-            print(f'alien file does not exist {alien_file}')
+            print('alien file does not exist ', alien_file)
     else:
         print('alien_file parameter not configured')
     return data
