+++ cohere_core/lib/cplib.py
@@ -2,9 +2,7 @@ from cohere_core.lib.cohlib import cohlib
 import cupy as cp
 import numpy as np
 import cupyx.scipy.ndimage
-import cupyx.scipy.stats as stats
-import cupyx.scipy.ndimage as sc
-
+import cupyx.scipy.signal
 
 class cplib(cohlib):
     def array(obj):
@@ -30,13 +28,10 @@ class cplib(cohlib):
         cp.save(filename, arr)
 
     def load(filename):
-        return cp.load(filename, allow_pickle=True)
+        return cp.load(filename)
 
     def dtype(arr):
         return arr.dtype
-
-    def astype(arr, dtype):
-        return arr.astype(dtype=dtype)
 
     def size(arr):
         return arr.size
@@ -55,36 +50,25 @@ class cplib(cohlib):
         rs = cp.random.RandomState(seed=seed)
         return cp.random.random(shape, dtype=cp.float32) + 1j * cp.random.random(shape, dtype=cp.float32)
 
-    def roll(arr, sft, axis=None):
-        if axis is None:
-            axis = list(range(len(sft)))
-        if type(sft) != list:
-            sft = [sft]
-        sft = [int(s) for s in sft]
-        return cp.roll(arr, sft, axis=axis)
-
-    def shift(arr, sft):
-        return sc.fourier_shift(arr, sft)
-
     def fftshift(arr):
         return cp.fft.fftshift(arr)
 
     def ifftshift(arr):
         return cp.fft.ifftshift(arr)
 
-    def fft(arr, norm='forward'):
-        return cp.fft.fftn(arr, norm=norm)
+    def shift(arr, sft, axis=None):
+        sft = [int(s) for s in sft]
+        return cp.roll(arr, sft, axis=axis)
 
-    def ifft(arr, norm='forward'):
-        return cp.fft.ifftn(arr, norm=norm)
+    def fft(arr):
+        return cp.fft.fftn(arr, norm='forward')
+
+    def ifft(arr):
+        return cp.fft.ifftn(arr, norm='forward')
 
     def fftconvolve(arr1, arr2):
-        return sc.convolve(arr1, arr2)
-        # return cupyx.scipy.signal.convolve(arr1, arr2, mode='same')
-
-    def correlate(arr1, arr2, mode='same', method='fft'):
-        from cupyx.scipy.signal import correlate
-        return correlate(arr1, arr2, mode, method)
+        return cupyx.scipy.ndimage.convolve(arr1, arr2)
+        # return cupyx.scipy.signal.aoconvolve(arr1, arr2, mode='same')
 
     def where(cond, x, y):
         return cp.where(cond, x, y)
@@ -104,8 +88,8 @@ class cplib(cohlib):
 
     def sum(arr, axis=None):
         sm = cp.sum(arr, axis)
-        # if axis is None:
-        #     return sm.tolist()
+        if axis is None:
+            return sm.tolist()
         return sm
 
     def real(arr):
@@ -164,15 +148,14 @@ class cplib(cohlib):
         inarr = cp.zeros((n_el))
         inarr[int(n_el / 2)] = 1.0
         inarr = cp.reshape(inarr, shape)
-        gaussian = sc.gaussian_filter(inarr, sigma)
+        gaussian = cupyx.scipy.ndimage.gaussian_filter(inarr, sigma)
         return gaussian / cp.sum(gaussian)
 
     def gaussian_filter(arr, sigma, **kwargs):
-        return sc.gaussian_filter(arr, sigma)
+        return cupyx.scipy.ndimage.gaussian_filter(arr, sigma)
 
     def center_of_mass(inarr):
-        t = sc.center_of_mass(cp.absolute(inarr))
-        return t
+        return cupyx.scipy.ndimage.center_of_mass(cp.absolute(inarr))
 
     def meshgrid(*xi):
         return cp.meshgrid(*xi)
@@ -182,6 +165,15 @@ class cplib(cohlib):
 
     def conj(arr):
         return cp.conj(arr)
+
+    def cos(arr):
+        return cp.cos(arr)
+
+    def linspace(start, stop, steps):
+        return cp.linspace(start, stop, steps)
+
+    def diff(arr, axis=None, prepend=0):
+        return cp.diff(arr, axis=axis, prepend=prepend)
 
     def array_equal(arr1, arr2):
         return cp.array_equal(arr1, arr2)
@@ -194,62 +186,3 @@ class cplib(cohlib):
 
     def clip(arr, min, max=None):
         return cp.clip(arr, min, max)
-
-    def diff(arr, axis=None, prepend=0):
-        return cp.diff(arr, axis=axis, prepend=prepend)
-
-    def gradient(arr, dx=1):
-        return cp.gradient(arr, dx)
-
-    def argmin(arr, axis=None):
-        return cp.argmin(arr, axis)
-
-    def take_along_axis(a, indices, axis):
-        return cp.take_along_axis(a, indices, axis)
-
-    def moveaxis(arr, source, dest):
-        return cp.moveaxis(arr, source, dest)
-
-    def lstsq(A, B):
-        return cp.linalg.lstsq(A, B, rcond=None)
-
-    def zeros(shape):
-        return cp.zeros(shape)
-
-    def indices(dims):
-        return cp.indices(dims)
-
-    def concatenate(tup, axis=0):
-        return cp.concatenate(tup, axis)
-
-    def amin(arr):
-        return cp.amin(arr)
-
-    def affine_transform(arr, matrix, order=3, offset=0):
-        return cupyx.scipy.ndimage.affine_transform(arr, matrix, order=order, offset=offset, prefilter=True)
-
-    def pad(arr, padding):
-        return cp.pad(arr, padding)
-
-    def histogram2d(meas, rec, n_bins=100, log=False):
-        norm = cp.max(meas) / cp.max(rec)
-        if log:
-            bins = cp.logspace(cp.log10(1), cp.log10(cp.max(meas)), n_bins+1)
-        else:
-            bins = n_bins
-        return cp.histogram2d(cp.ravel(meas), cp.ravel(norm*rec), bins)[0]
-
-    def calc_nmi(hgram):
-        h0 = stats.entropy(np.sum(hgram, axis=0))
-        h1 = stats.entropy(np.sum(hgram, axis=1))
-        h01 = stats.entropy(np.reshape(hgram, -1))
-        return (h0 + h1) / h01
-
-    def calc_ehd(hgram):
-        n = hgram.shape[0] * 1j
-        x, y = cp.mgrid[0:1:n, 0:1:n]
-        return cp.sum(hgram * cp.abs(x - y)) / cp.sum(hgram)
-
-    def clean_default_mem():
-        cp._default_memory_pool.free_all_blocks()
-        cp._default_pinned_memory_pool.free_all_blocks()
