+++ cohere_core/lib/torchlib.py
@@ -63,9 +63,6 @@ class torchlib(cohlib):
     def dtype(arr):
         return arr.dtype
 
-    def astype(arr, dtype):
-        return arr.astype(dtype=dtype)
-
     def size(arr):
         return torch.numel(arr)
 
@@ -76,26 +73,7 @@ class torchlib(cohlib):
         return arr.clone()
 
     def random(shape, **kwargs):
-        arr = torch.rand(shape, device=torchlib.device)
-        print(arr.dtype)
-        # return torch.rand(shape, device=torchlib.device)
-        return arr
-
-    def roll(arr, sft):
-        sft = [int(s) for s in sft]
-        dims = tuple([i for i in range(len(sft))])
-        try:
-            return torch.roll(arr, sft, dims)
-        except Exception as e:
-            print('not supported error: ' + repr(e))
-
-    def shift(arr, sft):
-        sft = [int(s) for s in sft]
-        dims = tuple([i for i in range(len(sft))])
-        try:
-            return torch.roll(arr, sft, dims)
-        except Exception as e:
-            print('not supported error: ' + repr(e))
+        return torch.rand(shape, device=torchlib.device)
 
     def fftshift(arr):
         # if roll is implemented before fftshift
@@ -114,6 +92,14 @@ class torchlib(cohlib):
         # return torch.roll(arr, shifts, dims=dims)
         try:
             return torch.fft.ifftshift(arr)
+        except Exception as e:
+            print('not supported error: ' + repr(e))
+
+    def shift(arr, sft):
+        sft = [int(s) for s in sft]
+        dims = tuple([i for i in range(len(sft))])
+        try:
+            return torch.roll(arr, sft, dims)
         except Exception as e:
             print('not supported error: ' + repr(e))
 
@@ -267,54 +253,6 @@ class torchlib(cohlib):
 
     def clip(arr, min, max=None):
         return torch.clip(arr, min, max)
-
-    def diff(arr, axis=None, prepend=0):
-        raise NotImplementedError
-
-    def gradient(arr, dx=1):
-        raise NotImplementedError
-
-    def argmin(arr, axis=None):
-        raise NotImplementedError
-
-    def take_along_axis(a, indices, axis):
-        raise NotImplementedError
-
-    def moveaxis(arr, source, dest):
-        raise NotImplementedError
-
-    def lstsq(A, B):
-        raise NotImplementedError
-
-    def zeros(shape):
-        raise NotImplementedError
-
-    def indices(dims):
-        raise NotImplementedError
-
-    def concatenate(tup, axis=0):
-        raise NotImplementedError
-
-    def amin(arr):
-        raise NotImplementedError
-
-    def affine_transform(arr, matrix, order=3, offset=0):
-        raise NotImplementedError
-
-    def pad(arr, padding):
-        raise NotImplementedError
-
-    def histogram2d(meas, rec, n_bins=100, log=False):
-        raise NotImplementedError
-
-    def calc_nmi(hgram):
-        raise NotImplementedError
-
-    def calc_ehd(hgram):
-        raise NotImplementedError
-
-    def clean_default_mem():
-        pass
 
 # a1 = torch.Tensor([0.1, 0.2, 0.3, 1.0, 1.2, 1.3])
 # a2 = torch.Tensor([10.1, 10.2, 10.3, 11.0])
