+++ cohere_core/utilities/dvc_utils.py
@@ -1,32 +1,4 @@
 import math
-import cohere_core.utilities.utils as ut
-
-_docformat__ = 'restructuredtext en'
-__all__ = ['set_lib',
-           'arr_property',
-           'pad_around',
-           'center_sync',
-           'gauss_conv_fft',
-           'shrink_wrap',
-           'shift_phase',
-           'zero_phase',
-           'conj_reflect',
-           'cross_correlation',
-           'check_get_conj_reflect',
-           'get_metric',
-           'all_metrics',
-           'dftups',
-           'dftregistration',
-           'register_3d_reconstruction',
-           'sub_pixel_shift',
-           'align_arrays_subpixel',
-           'align_arrays_pixel',
-           'fast_shift',
-           'correlation_err',
-           'remove_ramp',
-           'zero_phase_cc',
-           'breed'
-           ]
 
 
 def set_lib(dlib):
@@ -34,25 +6,22 @@ def set_lib(dlib):
     dvclib = dlib
 
 
-def arr_property(arr):
-    """
-    Used only in development. Prints max value of the array and max coordinates.
-
-    Parameters
-    ----------
-    arr : ndarray
-        array to find max
-    """
-    import numpy as np
-    arr1 = abs(arr)
-    print('norm', dvclib.sum(pow(abs(arr), 2)))
-    print('mean across all data', arr1.mean())
-    print('std all data', arr1.std())
-    print('sum',np.sum(arr))
-    # print('mean non zeros only', arr1[np.nonzero(arr1)].mean())
-    # print('std non zeros only', arr1[np.nonzero(arr1)].std())
-    max_coordinates = list(dvclib.unravel_index(dvclib.argmax(arr1), arr.shape))
-    print('max coords, value', max_coordinates, arr[max_coordinates[0], max_coordinates[1], max_coordinates[2]])
+def crop_center(arr, shape):
+    dims = dvclib.dims(arr)
+    principio = []
+    finem = []
+    for i in range(3):
+        principio.append(int((dims[i] - shape[i]) / 2))
+        finem.append(principio[i] + shape[i])
+    if len(shape) == 1:
+        cropped = arr[principio[0]: finem[0]]
+    elif len(shape) == 2:
+        cropped = arr[principio[0]: finem[0], principio[1]: finem[1]]
+    elif len(shape) == 3:
+        cropped = arr[principio[0]: finem[0], principio[1]: finem[1], principio[2]: finem[2]]
+    else:
+        raise NotImplementedError
+    return cropped
 
 
 def pad_around(arr, shape, val=0):
@@ -72,35 +41,6 @@ def pad_around(arr, shape, val=0):
     else:
         raise NotImplementedError
     return padded
-
-
-def center_sync(image, support):
-    """
-    Shifts the image and support arrays so the center of mass is in the center of array.
-    Parameters
-    ----------
-    image, support : ndarray, ndarray
-        image and support arrays to evaluate and shift
-    Returns
-    -------
-    image, support : ndarray, ndarray
-        shifted arrays
-    """
-    shape = image.shape
-    import numpy as np
-
-    # set center phase to zero, use as a reference
-    phi0 = math.atan2(image.flatten().imag[int(image.flatten().shape[0] / 2)],
-                   image.flatten().real[int(image.flatten().shape[0] / 2)])
-    image = image * dvclib.exp(-1j * phi0)
-
-    com = dvclib.center_of_mass(dvclib.astype(support, dtype=np.int32))
-    # place center of mass in the center
-    sft = [shape[i] / 2.0 - com[i] + .5  for i in range(len(shape))]
-    image = dvclib.roll(image, sft)
-    support =dvclib.roll(support, sft)
-
-    return image, support
 
 
 def gauss_conv_fft(arr, distribution):
@@ -126,8 +66,7 @@ def shrink_wrap(arr, threshold, sigma):
         subject array
 
     threshold : float
-        support is formed by points above this value
-    sigma: float
+        support is formed by points above this valuue
 
     sigmas : list
         sigmas in all dimensions
@@ -268,18 +207,14 @@ def get_metric(image, errs, metric_type):
         calculated metric for a given type
     """
     if metric_type == 'chi':
-        eval = errs[-1]
+        return errs[-1]
     elif metric_type == 'sharpness':
-        eval = dvclib.sum(dvclib.square(dvclib.square(dvclib.absolute(image))))
+        return dvclib.sum(dvclib.square(dvclib.square(dvclib.absolute(image))))
     elif metric_type == 'summed_phase':
         ph = shift_phase(image, 0)
-        eval = dvclib.sum(abs(ph))
+        return dvclib.sum(abs(ph))
     elif metric_type == 'area':
-        eval = dvclib.sum(shrink_wrap(image, .2, .5))
-
-    if type(eval) != float:
-        eval = eval.item()
-    return eval
+        return dvclib.sum(shrink_wrap(image, .2, .5))
 
 
 def all_metrics(image, errs):
@@ -297,23 +232,10 @@ def all_metrics(image, errs):
         calculated metrics for all types
     """
     metrics = {}
-    eval = errs[-1]
-    if type(eval) != float:
-        eval = eval.item()
-    metrics['chi'] = eval
-    eval = dvclib.sum(dvclib.square(dvclib.square(dvclib.absolute(image))))
-    if type(eval) != float:
-        eval = eval.item()
-    metrics['sharpness'] = eval
-    eval = dvclib.sum(abs(shift_phase(image, 0)))
-    if type(eval) != float:
-        eval = eval.item()
-    metrics['summed_phase'] = eval
-    eval = dvclib.sum(shrink_wrap(image, .2, .5))
-    if type(eval) != float:
-        eval = eval.item()
-    metrics['area'] = eval
-    
+    metrics['chi'] = errs[-1]
+    metrics['sharpness'] = dvclib.sum(dvclib.square(dvclib.square(dvclib.absolute(image))))
+    metrics['summed_phase'] = dvclib.sum(abs(shift_phase(image, 0)))
+    metrics['area'] = dvclib.sum(shrink_wrap(image, .2, .5))
     return metrics
 
 
@@ -395,7 +317,7 @@ def dftregistration(ref_arr, arr, usfac=2):
     c_c = pad_around(dvclib.fftshift(ref_arr) * dvclib.conj(dvclib.fftshift(arr)), large_shape, val=0j)
 
     # Compute crosscorrelation and locate the peak
-    c_c = dvclib.ifft(dvclib.ifftshift(c_c))
+    c_c = dvclib.ifft(dvclib.ifftshift(c_c))    #TODO will not work for aflib
     max_coord = dvclib.unravel_index(dvclib.argmax(c_c), dvclib.dims(c_c))
 
     if max_coord[0] > shape[0]:
@@ -449,9 +371,9 @@ def register_3d_reconstruction(ref_arr, arr):
     shift_2, shift_1, shift_0 : float, float
         pixel shifts between images
     """
-    r_shift_2, c_shift_2 = dftregistration(dvclib.fft(dvclib.sum(ref_arr, 2)), dvclib.fft(dvclib.sum(arr, 2)), 4)
-    r_shift_1, c_shift_1 = dftregistration(dvclib.fft(dvclib.sum(ref_arr, 1)), dvclib.fft(dvclib.sum(arr, 1)), 4)
-    r_shift_0, c_shift_0 = dftregistration(dvclib.fft(dvclib.sum(ref_arr, 0)), dvclib.fft(dvclib.sum(arr, 0)), 4)
+    r_shift_2, c_shift_2 = dftregistration(dvclib.fft(dvclib.sum(ref_arr, 2)), dvclib.fft(dvclib.sum(arr, 2)), 100)
+    r_shift_1, c_shift_1 = dftregistration(dvclib.fft(dvclib.sum(ref_arr, 1)), dvclib.fft(dvclib.sum(arr, 1)), 100)
+    r_shift_0, c_shift_0 = dftregistration(dvclib.fft(dvclib.sum(ref_arr, 0)), dvclib.fft(dvclib.sum(arr, 0)), 100)
 
     shift_2 = sum([r_shift_2, r_shift_1]) * 0.5
     shift_1 = sum([c_shift_2, r_shift_0]) * 0.5
@@ -489,7 +411,7 @@ def sub_pixel_shift(arr, row_shift, col_shift, z_shift):
     return dvclib.ifft(Greg)
 
 
-def align_arrays_subpixel(ref_arr, arr):
+def align_arrays(ref_arr, arr):
     """
     Shifts array to align with reference array.
 
@@ -508,126 +430,6 @@ def align_arrays_subpixel(ref_arr, arr):
     """
     (shift_2, shift_1, shift_0) = register_3d_reconstruction(abs(ref_arr), abs(arr))
     return sub_pixel_shift(arr, shift_2, shift_1, shift_0)
-
-
-def remove_ramp(arr, ups=1):
-    """
-    Smooths image array (removes ramp) by applaying math formula.
-    Parameters
-    ----------
-    arr : ndarray
-        array to remove ramp
-    ups : int
-        upsample factor
-    Returns
-    -------
-    ramp_removed : ndarray
-        smoothed array
-    """
-    # the remove ramp is called (now) by visualization when arrays are on cpu
-    # thus using functions from utilities will be ok for now
-    import cohere_core.utilities.utils as ut
-    shape = arr.shape
-    new_shape = [dim * ups for dim in shape]
-    padded = ut.pad_center(arr, new_shape)
-    padded_f = dvclib.fftshift(dvclib.fft(dvclib.ifftshift(padded)))
-    com = dvclib.center_of_mass(dvclib.square(dvclib.absolute(padded_f)))
-    sft = [new_shape[i] / 2.0 - com[i] + .5 for i in range(len(shape))]
-    sub_pixel_shifted = sub_pixel_shift(padded_f, *sft)
-    ramp_removed_padded = dvclib.fftshift(dvclib.ifft(dvclib.fftshift(sub_pixel_shifted)))
-    ramp_removed = ut.crop_center(ramp_removed_padded, arr.shape)
-
-    return ramp_removed
-
-
-# supposedly this is faster than np.roll or scipy interpolation shift.
-# https://stackoverflow.com/questions/30399534/shift-elements-in-a-numpy-array
-def fast_shift(arr, shifty, fill_val=0):
-    """
-    Shifts array by the shifty parameter.
-    Parameters
-    ----------
-    arr : ndarray
-        array to shift
-    shifty : ndarray
-        an array of integers/shifts in each dimension
-    fill_val : float
-
-    Returns
-    -------
-    ndarray, shifted array
-    """
-    dims = arr.shape
-    result = dvclib.full(dims, 1.0)
-    result *= fill_val
-    result_slices = []
-    arr_slices = []
-    for n in range(len(dims)):
-        if shifty[n] > 0:
-            result_slices.append(slice(shifty[n], dims[n]))
-            arr_slices.append(slice(0, -shifty[n]))
-        elif shifty[n] < 0:
-            result_slices.append(slice(0, shifty[n]))
-            arr_slices.append(slice(-shifty[n], dims[n]))
-        else:
-            result_slices.append(slice(0, dims[n]))
-            arr_slices.append(slice(0, dims[n]))
-    result_slices = tuple(result_slices)
-    arr_slices = tuple(arr_slices)
-    result[result_slices] = arr[arr_slices]
-    return result
-
-
-def align_arrays_pixel(ref, arr):
-    """
-    Aligns two arrays of the same dimensions with the pixel resolution. Used in reciprocal space.
-    Parameters
-    ----------
-    ref : ndarray
-        array to align with
-    arr : ndarray
-        array to align
-
-    Returns
-    -------
-    ndarray : aligned array
-    """
-    CC = dvclib.correlate(ref, arr, mode='same', method='fft')
-    CC_shifted = dvclib.ifftshift(CC)
-    shape = dvclib.array(CC_shifted.shape)
-    amp = dvclib.absolute(CC_shifted)
-    shift = dvclib.unravel_index(amp.argmax(), shape)
-    if dvclib.sum(shift) == 0:
-        return arr
-    intshift = dvclib.array(shift)
-    pixelshift = dvclib.where(intshift >= shape / 2, intshift - shape, intshift)
-    shifted_arr = fast_shift(arr, pixelshift)
-    return shifted_arr
-
-
-def correlation_err(refarr, arr):
-    """
-    author: Paul Frosik
-    The method is based on "Invariant error metrics for image reconstruction" by J. R. Fienup.
-    Returns correlation error between two arrays.
-
-    Parameters
-    ----------
-    refarr : ndarray
-        reference array
-    arr : ndarray
-        array to measure likeness with reference array
-
-    Returns
-    -------
-    float, correlation error
-    """
-    CC = dvclib.correlate(refarr, arr, mode='same', method='fft')
-    CCmax = CC.max()
-    rfzero = dvclib.sum(dvclib.absolute(refarr) ** 2)
-    rgzero = dvclib.sum(dvclib.absolute(arr) ** 2)
-    e = abs(1 - CCmax * dvclib.conj(CCmax) / (rfzero * rgzero)) ** 0.5
-    return  e
 
 
 def zero_phase_cc(arr1, arr2):
@@ -670,7 +472,7 @@ def breed(breed_mode, alpha_dir, image):
 
     """
     # load alpha from alpha dir
-    alpha = dvclib.load(ut.join(alpha_dir, 'image.npy'))
+    alpha = dvclib.load(alpha_dir + '/image.npy')
     if dvclib.array_equal(image, alpha):
         # it is alpha, no breeding
         return zero_phase(image)
@@ -680,7 +482,7 @@ def breed(breed_mode, alpha_dir, image):
     beta = image
     beta = zero_phase(beta)
     alpha = check_get_conj_reflect(beta, alpha)
-    alpha = align_arrays_subpixel(beta, alpha)
+    alpha = align_arrays(beta, alpha)
     alpha = zero_phase(alpha)
     ph_alpha = dvclib.angle(alpha)
     beta = zero_phase_cc(beta, alpha)
