+++ docs/source/config_main.rst
@@ -41,28 +41,21 @@ Parameters
     multipeak = True
 
 - separate_scans:
-| optional, deafults to False. In typical scenario the data from all scans in experiment are combined. If specified as separate scans, each scan will be processed separately and will have sub-experiment name containing scan index ex. "scan_9", where 9 is scan index
+| in typical scenario the data from all scans in experiment are combined. If specified as separate scans, each scan will be processed separately and will have sub-experiment name containing scan index ex. "scan_9", where 9 is scan index
 | example:
 ::
 
    separate_scans = False
 
 - separate_scan_ranges:
-| optional, defaults to False. In typical scenario the data from all scans in experiment are combined. If specified as separate scan ranges, each scan or scan range in the experiment will be processed separately and will have sub-experiment name containing scan index ex. "scan_9", where 9 is scan index, or "scan_10-15", where 10-15 is the scan range. The scans and scan ranges are defined in main configuration "config" file as scan parameter, and are part of experiment name.
+| in typical scenario the data from all scans in experiment are combined. If specified as separate scan ranges, each scan or scan range in the experiment will be processed separately and will have sub-experiment name containing scan index ex. "scan_9", where 9 is scan index, or "scan_10-15", where 10-15 is the scan range. The scans and scan ranges are defined in main configuration "config" file as scan parameter, and are part of experiment name.
 | example:
 ::
 
    separate_scan_ranges = True
 
-- auto-data:
-| optional, boolean parameter indicating automated data preprocessing, which includes exclusion of outlier scans in multi-scan data and auto calculation of intensity threshold. Defaults to False.
-| example:
-::
-
-    multipeak = True
-
 - converter_ver:
-| mandatory after ver 3.0, if not configured, it will be auto-updated by scripts to match the latest converter version.
+| optional, if not configured, it will be auto-updated by scripts to match the latest converter version.
 | example:
 ::
